requirements.txt
python-docx>=1.2.0
PyYAML>=6.0
python-frontmatter>=1.1.0
